+++ scripts/macros/2014/spells/tollTheDead.js
@@ -1,7 +1,7 @@
 import {activityUtils, genericUtils, itemUtils} from '../../../utils.js';
 async function use({trigger, workflow}) {
     if (!workflow.targets.size) return;
-    let validTargets = workflow.targets.filter(i => i.actor.system.attributes.hp.value < i.actor.system.attributes.hp.max);
+    let validTargets = workflow.targets.filter(i => i.actor.system.attributes.hp.value < i.actor.system.attributes.hp.effectiveMax);
     if (!validTargets.size) return;
     let formula = itemUtils.getConfig(workflow.item, 'formula');
     let activityData = activityUtils.withChangedDamage(workflow.activity, formula);