+++ scripts/macros/2024/classFeatures/cleric/life/preserveLife.js
@@ -31,15 +31,8 @@ async function use({trigger: {entity: item}, workflow}) {
         if (playAnimation) new Sequence().effect().atLocation(target).file('jb2a.cure_wounds.400px.blue').play();
     }
 }
-async function added({trigger: {entity: item, actor}}) {
-    let channelDivinity = itemUtils.getItemByIdentifier(actor, 'channelDivinity');
-    if (!channelDivinity) return;
-    let activity = activityUtils.getActivityByIdentifier(item, 'use', {strict: true});
-    if (!activity) return;
-    let itemData = genericUtils.duplicate(item.toObject());
-    itemData.system.activities[activity.id].consumption.targets[0].target = channelDivinity.id;
-    let path = 'system.activities.' + activity.id + '.consumption.targets';
-    await genericUtils.update(item, {[path]: itemData.system.activities[activity.id].consumption.targets});
+async function added({trigger: {entity: item}}) {
+    await itemUtils.correctActivityItemConsumption(item, ['use'], 'channelDivinity');
 }
 export let preserveLife = {
     name: 'Preserve Life',